Fix ricoh camlib compilation warnings.
* g3.c:
  * get_file_func(...): Change variable 'seek' type to 'int'. Currently 'sscanf' already reads a signed decimal into 'seek' and a following comparison uses negative constant (-2). This change should not change behavior in practice.
  * folder_list_func(...): Variable 'n' is used in trivial 'for' loop already as unsigned so change the the type accordingly. No behavior changes.
  * file_list_func(...): Same as above.
* library.c:
  * camera_init(...): Add cast to 'unsigned int' to avoid signed-unsigned compare warning. This does not change code behavior.

+++ camlibs/ricoh/library.c
@@ -565,7 +565,7 @@ camera_init (Camera *camera, GPContext *context)
 	}
 
 	/* Contact made. Do we need to change the speed? */
-	if (settings.serial.speed != speed) {
+	if ((unsigned int)settings.serial.speed != speed) {
 		for (i = 0; speeds[i].speed; i++)
 			if (speeds[i].speed == speed)
 				break;
